Extract stop action
