+++ src/jlmkr/donor/jlmkr.py
@@ -751,26 +751,7 @@ def edit_jail(jail_name):
     return 0
 
 
-def stop_jail(jail_name):
-    """
-    Stop jail with given name and wait until stopped.
-    """
-
-    if not jail_is_running(jail_name):
-        return 0
-
-    returncode = subprocess.run(["machinectl", "poweroff", jail_name]).returncode
-    if returncode != 0:
-        eprint("Error while stopping jail.")
-        return returncode
-
-    print(f"Wait for {jail_name} to stop", end="", flush=True)
-
-    while jail_is_running(jail_name):
-        time.sleep(1)
-        print(".", end="", flush=True)
-
-    return 0
+from actions.stop import stop_jail
 
 
 def remove_jail(jail_name):
